Added operations to statements

+++ Tests Resources/expressions/while.java
@@ -2,7 +2,7 @@ while ( true ){
   int x =1;
 }
 while ( true ){
-  int x =1;
+  float x =1;
 }
 while ( true ){
   int x =1;
